+++ examples/multiFileProject/multiFileProject.cpp
@@ -1,0 +1,15 @@
+/****************************************************************************************************************************
+  multiFileProject.cpp
+  For ESP8266, ESP32 and STM32 with built-in LAN8742A Ethernet (Nucleo-144, DISCOVERY, etc)
+
+  AsyncHTTPSRequest_Generic is a library for the ESP8266, ESP32 and currently STM32 run built-in Ethernet WebServer
+
+  Based on and modified from AsyncHTTPSRequest Library (https://github.com/boblemaire/AsyncHTTPSRequest)
+
+  Built by Khoi Hoang https://github.com/khoih-prog/AsyncHTTPSRequest_Generic
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#include "multiFileProject.h"