+++ examples/multiFileProject/multiFileProject.h
@@ -1,0 +1,18 @@
+/****************************************************************************************************************************
+  multiFileProject.h
+  For ESP8266, ESP32 and STM32 with built-in LAN8742A Ethernet (Nucleo-144, DISCOVERY, etc)
+
+  AsyncHTTPSRequest_Generic is a library for the ESP8266, ESP32 and currently STM32 run built-in Ethernet WebServer
+
+  Based on and modified from AsyncHTTPSRequest Library (https://github.com/boblemaire/AsyncHTTPSRequest)
+
+  Built by Khoi Hoang https://github.com/khoih-prog/AsyncHTTPSRequest_Generic
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#pragma once
+
+// Can be included as many times as necessary, without `Multiple Definitions` Linker Error
+#include "AsyncHTTPSRequest_Generic.hpp"