+++ src/AsyncHTTPSRequest_Impl_Generic.h
@@ -16,7 +16,7 @@
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
   You should have received a copy of the GNU General Public License along with this program.  If not, see <https://www.gnu.org/licenses/>.  
  
-  Version: 1.2.0
+  Version: 1.3.0
   
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -24,12 +24,15 @@
   1.1.0    K Hoang     23/10/2021 Add support to ESP32-based WT32-ETH01 using LAN8720
   1.1.1    K Hoang     29/11/2021 Auto detect ESP32 core version and improve connection time for WT32_ETH01
   1.2.0    K Hoang     30/12/2021 Fix `multiple-definitions` linker error
+  1.3.0    K Hoang     23/01/2022 Enable compatibility with old code to include only AsyncHTTPSRequest_Generic.h
  *****************************************************************************************************************************/
  
 #pragma once
 
 #ifndef ASYNC_HTTPS_REQUEST_GENERIC_IMPL_H
 #define ASYNC_HTTPS_REQUEST_GENERIC_IMPL_H
+
+#include <AsyncTCP_SSL.h>
 
 #define CANT_SEND_BAD_REQUEST       F("Can't send() bad request")